+++ www/controllers/tasks.controller.php
@@ -118,7 +118,7 @@ class TasksController extends Controller
 			//TODO : ajouter la variable $dataTableConf en paramètre du hook => résoudre le souci de "order"
 			//      $dataTableConf = array(
 			//          'language' => array(
-			//              'url' => $context->getRootUrl() . 'vendor/data-tables/french.json',
+			//              'url' => $context->getControllerUrl() . 'vendor/data-tables/french.json',
 			//          ),
 			//          'order' => array(),
 			//          'responsive' => true,
@@ -303,7 +303,7 @@ class TasksController extends Controller
 					//$("#expedition-list").DataTable(<?php //echo json_encode($dataTableConf) ?>//);
 					$("#projettask-list").DataTable({
 						"language": {
-							"url": "<?php print $context->getRootUrl(); ?>vendor/data-tables/french.json"
+							"url": "<?php print $context->getControllerUrl(); ?>vendor/data-tables/french.json"
 						},
 						//"order": [[<?php echo ($total_more_fields + 2); ?>, 'desc']],