+++ www/tpl/header.tpl.php
@@ -28,52 +28,52 @@ $metaTitle.= !empty($conf->global->EACCESS_TITLE)?$conf->global->EACCESS_TITLE:$
     <title><?php echo dol_htmlentities($metaTitle); ?></title>
 
     <!-- Bootstrap core CSS -->
-    <link href="<?php print $context->getRootUrl(); ?>vendor/bootstrap/css/bootstrap.css" rel="stylesheet">
+    <link href="<?php print $context->getControllerUrl(); ?>vendor/bootstrap/css/bootstrap.css" rel="stylesheet">
 
     <!-- Custom fonts for this template -->
-    <link href="<?php print $context->getRootUrl(); ?>vendor/font-awesome/css/font-awesome.min.css" rel="stylesheet" type="text/css">
+    <link href="<?php print $context->getControllerUrl(); ?>vendor/font-awesome/css/font-awesome.min.css" rel="stylesheet" type="text/css">
     <link href='https://fonts.googleapis.com/css?family=Open+Sans:300italic,400italic,600italic,700italic,800italic,400,300,600,700,800' rel='stylesheet' type='text/css'>
 
 
     <!-- Plugin CSS -->
-    <link href="<?php print $context->getRootUrl(); ?>vendor/magnific-popup/magnific-popup.css" rel="stylesheet">
+    <link href="<?php print $context->getControllerUrl(); ?>vendor/magnific-popup/magnific-popup.css" rel="stylesheet">
 
     <!-- Custom styles for this template -->
-    <link href="<?php print $context->getRootUrl(); ?>css/creative.css.php" rel="stylesheet">
+    <link href="<?php print $context->getControllerUrl(); ?>css/creative.css.php" rel="stylesheet">
 
-    <link rel="stylesheet" type="text/css" href="<?php print $context->getRootUrl(); ?>vendor/data-tables/datatables.min.css"/>
+    <link rel="stylesheet" type="text/css" href="<?php print $context->getControllerUrl(); ?>vendor/data-tables/datatables.min.css"/>
 
 
     <!-- Bootstrap core JavaScript -->
-    <script src="<?php print $context->getRootUrl(); ?>vendor/jquery/jquery.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/jquery/jquery.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/bootstrap/js/bootstrap.bundle.min.js"></script>
 
 	<!-- bootbox code -->
-	<script src="<?php print $context->getRootUrl(); ?>vendor/bootbox/bootbox.all.min.js"></script>
+	<script src="<?php print $context->getControllerUrl(); ?>vendor/bootbox/bootbox.all.min.js"></script>
 
     <!-- Plugin JavaScript -->
-    <script src="<?php print $context->getRootUrl(); ?>vendor/jquery-easing/jquery.easing.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/scrollreveal/scrollreveal.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/magnific-popup/jquery.magnific-popup.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/data-tables/datatables.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/data-tables/jquery.dataTables.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/data-tables/dataTables.bootstrap4.min.js"></script>
-    <script src="<?php print $context->getRootUrl(); ?>vendor/data-tables/Buttons-1.5.1/js/buttons.print.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/jquery-easing/jquery.easing.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/scrollreveal/scrollreveal.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/magnific-popup/jquery.magnific-popup.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/data-tables/datatables.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/data-tables/jquery.dataTables.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/data-tables/dataTables.bootstrap4.min.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/data-tables/Buttons-1.5.1/js/buttons.print.js"></script>
 
     <!-- Plugin Notify -->
-    <script src="<?php print $context->getRootUrl(); ?>vendor/noty/noty.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="<?php print $context->getRootUrl(); ?>vendor/noty/noty.css"/>
-    <link rel="stylesheet" type="text/css" href="<?php print $context->getRootUrl(); ?>vendor/noty/themes/metroui.css"/>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/noty/noty.min.js"></script>
+    <link rel="stylesheet" type="text/css" href="<?php print $context->getControllerUrl(); ?>vendor/noty/noty.css"/>
+    <link rel="stylesheet" type="text/css" href="<?php print $context->getControllerUrl(); ?>vendor/noty/themes/metroui.css"/>
 
     <!-- Plugin Select -->
-    <script src="<?php print $context->getRootUrl(); ?>vendor/bootstrap-select/dist/js/bootstrap-select.min.js"></script>
-    <link rel="stylesheet" type="text/css" href="<?php print $context->getRootUrl(); ?>vendor/bootstrap-select/dist/css/bootstrap-select.min.css"/>
+    <script src="<?php print $context->getControllerUrl(); ?>vendor/bootstrap-select/dist/js/bootstrap-select.min.js"></script>
+    <link rel="stylesheet" type="text/css" href="<?php print $context->getControllerUrl(); ?>vendor/bootstrap-select/dist/css/bootstrap-select.min.css"/>
 
     <!-- Custom scripts for this template -->
-    <script src="<?php print $context->getRootUrl(); ?>js/creative.js"></script>
+    <script src="<?php print $context->getControllerUrl(); ?>js/creative.js"></script>
 
 	<!-- Custom include ckeditor -->
-	<script src="<?php print $context->getRootUrl(); ?>vendor/ckeditor/ckeditor.js"></script>
+	<script src="<?php print $context->getControllerUrl(); ?>vendor/ckeditor/ckeditor.js"></script>
 
 <?php
 
@@ -88,7 +88,7 @@ $metaTitle.= !empty($conf->global->EACCESS_TITLE)?$conf->global->EACCESS_TITLE:$
     }
 
     // Mobile appli like icon
-    print '	<link rel="manifest" href="'.$context->getRootUrl().'manifest.json.php'.'" />' . "\r\n";
+    print '	<link rel="manifest" href="'.$context->getControllerUrl().'manifest.json.php'.'" />' . "\r\n";
 
 	$primaryColor = !empty($conf->global->EACCESS_PRIMARY_COLOR)?$conf->global->EACCESS_PRIMARY_COLOR:'#F05F40';
 	$backgroundColor = !empty($conf->global->EACCESS_APPLI_COLOR)?$conf->global->EACCESS_APPLI_COLOR:$primaryColor;